move inc into deps
